Make cached AliBi matrix static
This significantly reduces the memory cost of caching the AliBi matrix in deep networks. The previous version computed a new matrix for each attention layer. For a 2.7B GPT-2 style model, that means storing this matrix 32 times, which becomes prohibitively expensive when the sequence lengths get past 2,048 since this matrix grows with `O(length^2)`.

A potential downside of this change is that it ties the shape across all attention layers. That should be fine for decoder-only models, but may become a problem for encoder-decoder models where the cross-attention matrix may need to be rectangular instead of square. On the other hand, since the cached matrix always grows to the largest (key) sequence length seen and uses slicing to match the actual input dimensions, this might not be a problem.

Note that there is still one replica of the cached matrix on each accelerator.